Restructure frontend codebase
