+++ frontend/Makefile.toml
@@ -1,3 +1,10 @@
-[tasks.build]
+[tasks.build-trunk]
 command = "trunk"
 args = ["build", "--public-url", "/static/frontend"]
+
+[tasks.build-tailwindcss]
+command = "npx"
+args = ["--yes", "tailwindcss", "-i", "./styles/app.css", "-o", "../static/styles/app.css"]
+
+[tasks.build]
+dependencies = ["build-trunk", "build-tailwindcss"]
