+++ frontend/src/main.rs
@@ -1,8 +1,7 @@
 use app::App;
 
 mod app;
-mod bodies;
-mod responses;
+mod net;
 
 fn main() {
     wasm_logger::init(wasm_logger::Config::default());
