+++ frontend/src/net/mod.rs
@@ -1,0 +1,2 @@
+pub mod bodies;
+pub mod responses;
